Revert "[jtx rewrite] Merge `GeoFenceRadiusHandler` into `GeoHandler`"
This reverts commit 140ec4749de0009365e93f24512356d46cad8f8b.
