+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/jtx/handler/GeoFenceRadiusHandler.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
+ */
+
+package at.bitfire.synctools.mapping.jtx.handler
+
+import android.content.Entity
+import at.bitfire.synctools.icalendar.plusAssign
+import at.bitfire.synctools.mapping.jtx.JtxProperty
+import at.techbee.jtx.JtxContract
+import net.fortuna.ical4j.model.component.CalendarComponent
+import net.fortuna.ical4j.model.property.XProperty
+
+class GeoFenceRadiusHandler : JtxObjectEntityHandler {
+    override fun process(from: Entity, main: Entity, to: CalendarComponent) {
+        from.entityValues.getAsInteger(JtxContract.JtxICalObject.GEOFENCE_RADIUS)?.let { radius ->
+            to += XProperty(JtxProperty.X_GEOFENCE_RADIUS, radius.toString())
+        }
+    }
+}
